requirements.txt
fastapi~=0.115.5
dishka==1.4.2
python-jose==3.3.0
bcrypt==4.2.1
pydantic==2.10.1
pydantic-settings==2.6.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
typer==0.14.0
json-logging==1.3.0
starlette-context==0.3.6
aiosmtplib==3.0.2
